# employee/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/employees/employee/public/",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get all employees",
                "operationId": "App\\Http\\Controllers\\EmployeeController::employeesList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit of the item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Add a new employee      *Note: 1:male, 2:female, 3:other",
                "operationId": "App\\Http\\Controllers\\EmployeeController::createEmployee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "integar"
                                    },
                                    "birth_date": {
                                        "type": "date"
                                    },
                                    "hire_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "first_name": "Mustafa",
                                    "last_name": "Dahab",
                                    "gender": 1,
                                    "birth_date": "1988-11-11",
                                    "hire_date": "2020-10-10"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/employees/top_paid_employees": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get Top paid employees",
                "operationId": "App\\Http\\Controllers\\EmployeeController::employeesTopPaid",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit of the item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/employees/employees_salary_by_age": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get Average employees Salary by Age",
                "operationId": "App\\Http\\Controllers\\EmployeeController::employeesArgSalary",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit of the item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/employees/{id}": {
            "put": {
                "tags": [
                    "Employee"
                ],
                "summary": "Update employee",
                "operationId": "App\\Http\\Controllers\\EmployeeController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "integar"
                                    },
                                    "birth_date": {
                                        "type": "date"
                                    },
                                    "hire_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "first_name": "Mustafa",
                                    "last_name": "Dahab",
                                    "gender": 1,
                                    "birth_date": "1988-11-11",
                                    "hire_date": "2020-10-10"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Delete employee",
                "operationId": "App\\Http\\Controllers\\EmployeeController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}